Добавил новый класс Json логина
